refactor(dashboard): relocate isRecord to client-safe lib/utils; guard untrusted Telegram response
